+++ R/convertTranscriptsToGenes-methods.R
@@ -1,14 +1,15 @@
 #' @name convertTranscriptsToGenes
 #' @inherit bioverbs::convertTranscriptsToGenes
+#'
 #' @note For objects containing a count matrix, the object rows will be
 #'   collapsed to gene level using `aggregateRows`. This applies to our
 #'   `SummarizedExperiment` method.
 #' @note Updated 2019-07-28.
 #'
-#' @inheritParams params
+#' @inheritParams acidroxygen::params
 #' @param aggregate `logical(1)`.
-#'   For objects supporting [dim][base::dim], aggregate counts to gene level and
-#'   collapse the matrix.
+#'   For objects supporting [`dim()`][base::dim], aggregate counts to gene level
+#'   and collapse the matrix.
 #' @param ... Additional arguments.
 #'
 #' @return