+++ R/counts-methods.R
@@ -7,9 +7,9 @@
 #'
 #' @name counts
 #' @aliases counts<-
-#' @note Updated 2019-07-28.
+#' @note Updated 2019-08-06.
 #'
-#' @inheritParams params
+#' @inheritParams acidroxygen::params
 #' @param ... Additional arguments.
 #'
 #' @return Matrix.
@@ -42,12 +42,11 @@ NULL
 
 
 
-## Updated 2019-07-22.
+## Updated 2019-08-06.
 `counts,SummarizedExperiment` <-  # nolint
     function(object) {
         validObject(object)
-        assert(isSubset("counts", assayNames(object)))
-        assays(object)[["counts"]]
+        assay(object, i = "counts")
     }
 
 
@@ -62,11 +61,16 @@ setMethod(
 
 
 
-## Updated 2019-07-22.
-`counts<-,SummarizedExperiment,ANY` <-  # nolint
+## Updated 2019-08-06.
+`counts<-,SummarizedExperiment,matrix` <-  # nolint
     function(object, value) {
+        assert(
+            all(!is.na(value)),
+            all(is.finite(value)),
+            all(value >= 0L)
+        )
+        assay(object, i = "counts") <- value
         validObject(object)
-        assays(object)[["counts"]] <- value
         object
     }
 
@@ -74,11 +78,30 @@ setMethod(
 
 #' @rdname counts
 #' @export
-setMethod(
-    f = "counts<-",
+setReplaceMethod(
+    f = "counts",
     signature = signature(
         object = "SummarizedExperiment",
-        value = "ANY"
+        value = "matrix"
     ),
-    definition = `counts<-,SummarizedExperiment,ANY`
+    definition = `counts<-,SummarizedExperiment,matrix`
+)
+
+
+
+## Updated 2019-08-06.
+`counts<-,SummarizedExperiment,Matrix` <-  # nolint
+    `counts<-,SummarizedExperiment,matrix`
+
+
+
+#' @rdname counts
+#' @export
+setReplaceMethod(
+    f = "counts",
+    signature = signature(
+        object = "SummarizedExperiment",
+        value = "Matrix"
+    ),
+    definition = `counts<-,SummarizedExperiment,Matrix`
 )