+++ R/makeTx2Gene.R
@@ -7,7 +7,7 @@
 #' @name makeTx2Gene
 #' @note Updated 2019-07-28.
 #'
-#' @inheritParams params
+#' @inheritParams acidroxygen::params
 #'
 #' @return `Tx2Gene`.
 #'