+++ R/reexports.R
@@ -43,12 +43,98 @@ NULL
 
 
 
-## Assignment pipe
+#' @importFrom GenomicRanges GRanges
+#' @export
+GenomicRanges::GRanges
+
+#' @importFrom GenomicRanges GRangesList
+#' @export
+GenomicRanges::GRangesList
+
+#' @importFrom IRanges IRanges
+#' @export
+IRanges::IRanges
+
+#' @importFrom S4Vectors DataFrame
+#' @export
+S4Vectors::DataFrame
+
+#' @importFrom S4Vectors SimpleList
+#' @export
+S4Vectors::SimpleList
+
+#' @importFrom S4Vectors mcols
+#' @export
+S4Vectors::mcols
+
+#' @importFrom S4Vectors mcols<-
+#' @export
+S4Vectors::`mcols<-`
+
+#' @importFrom S4Vectors metadata
+#' @export
+S4Vectors::metadata
+
+#' @importFrom S4Vectors metadata<-
+#' @export
+S4Vectors::`metadata<-`
+
+#' @importFrom SingleCellExperiment reducedDims
+#' @export
+SingleCellExperiment::reducedDims
+
+#' @importFrom SingleCellExperiment reducedDims<-
+#' @export
+SingleCellExperiment::`reducedDims<-`
+
+#' @importFrom SummarizedExperiment assayNames
+#' @export
+SummarizedExperiment::assayNames
+
+#' @importFrom SummarizedExperiment assay
+#' @export
+SummarizedExperiment::assay
+
+#' @importFrom SummarizedExperiment assay<-
+#' @export
+SummarizedExperiment::`assay<-`
+
+#' @importFrom SummarizedExperiment assays
+#' @export
+SummarizedExperiment::assays
+
+#' @importFrom SummarizedExperiment assays<-
+#' @export
+SummarizedExperiment::`assays<-`
+
+#' @importFrom SummarizedExperiment colData
+#' @export
+SummarizedExperiment::colData
+
+#' @importFrom SummarizedExperiment colData<-
+#' @export
+SummarizedExperiment::`colData<-`
+
+#' @importFrom SummarizedExperiment rowData
+#' @export
+SummarizedExperiment::rowData
+
+#' @importFrom SummarizedExperiment rowData<-
+#' @export
+SummarizedExperiment::`rowData<-`
+
+#' @importFrom SummarizedExperiment rowRanges
+#' @export
+SummarizedExperiment::rowRanges
+
+#' @importFrom SummarizedExperiment rowRanges<-
+#' @export
+SummarizedExperiment::`rowRanges<-`
+
 #' @importFrom magrittr %<>%
 #' @export
 magrittr::`%<>%`
 
-## Exposition pipe
 #' @importFrom magrittr %$%
 magrittr::`%$%`
 
@@ -73,8 +159,6 @@ magrittr::set_names
 #' @importFrom magrittr set_rownames
 #' @export
 magrittr::set_rownames
-
-
 
 #' @importFrom tibble tibble
 #' @export