+++ R/sampleNames-methods.R
@@ -7,7 +7,7 @@
 #' @name sampleNames
 #' @note Updated 2019-07-28.
 #'
-#' @inheritParams params
+#' @inheritParams acidroxygen::params
 #' @param ... Additional arguments.
 #'
 #' @return Named `character`.