+++ tests/testthat/test-headtail.R
@@ -34,6 +34,6 @@ test_that("SummarizedExperiment", {
 })
 
 test_that("ASCII mode", {
-    output <- capture.output(headtail(rse, ascii = TRUE))
+    output <- capture.output(headtail(rse, unicode = FALSE))
     expect_true(grepl("...", output[[1L]]))
 })